Rewrite tap client connection logic
Instead of hooking into the memcached core, we implement everything
wiht a single thread within the engine. By doing so we can remove the
complex logic in the core, and remove everything related to -O from
memcached. The libevent use in the memcached core is intended for
use by the memcached core only, and not for providing extra sockets
to the engines...

Change-Id: I254db90b1c218d1e4373bc511c743b96ef78adf7